pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry]
authors = ["Cat Dev Group"]
description = ""
name = "bot"
version = "1.0.0"

[tool.poetry.dependencies]
python = "~3.9"
PyYAML = "~5.4"
"discord.py" = {url = "https://github.com/Rapptz/discord.py/archive/master.zip"}
loguru = "~0.5"
python-dotenv = "~0.17"

[tool.poetry.dev-dependencies]
black = "~20.8b1"
pytest = "~5.2"
taskipy = "~1.8.1"
isort = "~5.9.3"

[tool.taskipy.tasks]
start = "poetry run python -m bot"
format = "poetry run black . && poetry run isort ."